requirements.txt
pypdf>=4.0.0
PyMuPDF>=1.24.0
psycopg2-binary>=2.9.0
